services/ingest-service/requirements.txt
opencv-python-headless>=4.9.0
kafka-python>=2.0.2
prometheus-client>=0.20.0
